Implemented the NWS observation station fallback mechanism according to the plan. Here's what was accomplished:
Changes Made

  1. Database Schema (WeatherEntity.kt)
  - Added stationId: String? field to track which NWS observation station provided the data
  - Field is nullable and only populated for actual observations (not forecasts)

  2. Database Migration (WeatherDatabase.kt)
  - Updated database version from 7 to 8
  - Added MIGRATION_7_8 to add the stationId column
  - Migration registered in the database builder

  3. Station Caching (WeatherRepository.kt)
  - Added getCachedStations() function with 24-hour TTL
  - Added cacheStations() function to persist station lists in SharedPreferences
  - Reduces API calls from ~8/fetch to ~1/day for station list

  4. Fallback Logic (WeatherRepository.kt)
  - Changed fetchDayObservations() return type from Pair<Int, Int>? to Triple<Int, Int, String>? to include station ID
  - Implemented sequential retry through up to 5 stations when data is missing
  - Early termination when data is found (no unnecessary API calls)
  - Comprehensive logging at each step

  5. Station ID Persistence (WeatherRepository.kt)
  - Added stationByDate map to track station IDs during fetch
  - Updated WeatherEntity creation to include station ID from observations
  - Station ID is null for forecast data, populated only for actual observations

Verified Behavior
  From the logs, the system is working perfectly:

  ✅ Fallback is working:
  - First station (AW020) had no data for 2026-02-02
  - System automatically tried second station (KNUQ)
  - Successfully retrieved 30 observations from KNUQ

  ✅ Caching is working:
  - First fetch: "Fetching station list from API" (52 stations found)
  - Subsequent fetches: "Using cached stations (52 total)"

  ✅ Station tracking is working:
  - Logs show: "Station KNUQ provided data for 2026-02-02 (H:53 L:41) after 2 attempts"
  - Logs show: "Got observations for 2026-02-02 H=53 L=41 from station KNUQ"

  ✅ Early termination working:
  - When first station has data (AW020 for 2026-01-31), only 1 attempt is made
  - When first station fails (AW020 for 2026-02-02), system tries KNUQ and succeeds on attempt 2

  Performance Impact

  API Efficiency:
  - Station list: Fetched once, then cached for 24 hours (reduces ~23-47 calls/day)

+++ app/src/test/java/com/weatherwidget/util/TemperatureInterpolatorTest.kt
@@ -21,7 +21,7 @@ class TemperatureInterpolatorTest {
             dateTime = dateTime,
             locationLat = 37.42,
             locationLon = -122.08,
-            temperature = temp,
+            temperature = temp.toFloat(),
             source = "OPEN_METEO",
             fetchedAt = System.currentTimeMillis()
         )
@@ -47,7 +47,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 30)
         )
 
-        assertEquals(70, result)
+        assertEquals(70.0f, result)
     }
 
     @Test
@@ -61,7 +61,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 30)
         )
 
-        assertEquals(72, result)
+        assertEquals(72.0f, result)
     }
 
     @Test
@@ -77,7 +77,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 30)
         )
 
-        assertEquals(72, result)
+        assertEquals(72.0f, result)
     }
 
     @Test
@@ -93,7 +93,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 15)
         )
 
-        assertEquals(71, result)
+        assertEquals(71.0f, result)
     }
 
     @Test
@@ -109,7 +109,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 45)
         )
 
-        assertEquals(73, result)
+        assertEquals(73.0f, result)
     }
 
     @Test
@@ -125,7 +125,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 30)
         )
 
-        assertEquals(73, result)
+        assertEquals(73.0f, result)
     }
 
     @Test
@@ -141,7 +141,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 30)
         )
 
-        assertEquals(70, result)
+        assertEquals(70.0f, result)
     }
 
     @Test
@@ -157,7 +157,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 0)
         )
 
-        assertEquals(70, result)
+        assertEquals(70.0f, result)
     }
 
     @Test
@@ -173,7 +173,7 @@ class TemperatureInterpolatorTest {
             LocalDateTime.of(2026, 1, 15, 14, 0)
         )
 
-        assertEquals(68, result)
+        assertEquals(68.0f, result)
     }
 
     // Tests for getUpdatesPerHour
